Revert breaking edits and fix tests
It wasn't working before because the default redirectURI in the Config had been changed.
cli is a predefined client in arduino authentication system that requires a certain redirectURI.
Changing the redirectURI means you have to change the ClientID as well.

Now the tests are run providing -user and -pass flags, in order not to leak passwords when the arduino-cli will be made public.