Fixed double computation of install folder for libraries (WIP)
There are still bugs to be solved.

The zip coming from the repository is unpacked *as is* into libraries
directory, this lead to a path like LibName-1.0 to be extracted directly
from the .zip archive when, instead, it should be ignored.

The lib manager should create a sketchbook/libraries/LibName folder and
unzip the .zip archive there ignoring the root folder contained in the
archive itself.