Dramatically simplified visualization logic
This commit is quite big because visualization logic was deeply
tied with business logic. Trying to separete one piece of visualization
logic caused a chain reaction resulting in this massive refactoring.

Some of the issues resolved:

- the download is now handled by the `grab` library
- results visualization is greatly simplified by removing calls to
  the formatter package from the cores package (it wasn't clear why,
  for example, the ParallelDownload function should also handle
  formatting of the output)
- the formatter package has now a nice and handy progress bar formatter
- the package download subroutine was repeated inside the install
  subroutine. This duplication is now removed.
- classes are now passed instead of generic `interface{}`

Anyway the business logic is now much more simple and straighforward and
also the visualization had benefit from this change.