Don't scan hidden directories in arduino-builder
Arduino-builder scans the core to determine if core config files have
changed between invocations.  Unfortunately, it goes into SCCS dirs
such as .git (which can have 100,000+ files).  This is wasted effort
and can cause massive amounts of runtime and memory use when core
developers are building in their git clones.

Use a new function already added by the builder/utils.go to determine
if a file or folder is SCCS, and if so skip it in the rebuild-required
check.

Fixes https://github.com/arduino/arduino-builder/issues/327

Signed-off-by: Earle F. Philhower, III <earlephilhower@yahoo.com>