Replaced SourceFile.Origin field with SourceFile.Library
Previously the Origin interface{} field was used to store the "origin"
of the source file (being sketch or library). Actually Origin is really
needed only to reconstruct BuildPath and RootPath when the origin is a
Library, in all the other cases we can reconstruct BuildPath and RootPath
looking inside Context.

With the above in mind, this commit removes the non-idiomatic
interface{} by replacing it with a Library, so now the meaning of the
SourceFile.Library field is "the library where the source file is
contained or nil if not part of a library".