Add tests for ResolveFQBN
This adds tests for ResolveFQBN (which returns more details than the
FindBoardWithFQBN that was already tested), but also adds an extra
boards.txt with boards that are derived from another platform/core (i.e.
have `build.core=package:core`.

These new tests are currently failing, because of a bug and insufficient
error handling, which will be fixed in subsequent commits.