Fix resolution of referenced cores
A core is referenced by specifying `build.core=package:core` in board
properties. However, the code used the second part (core) both for
looking up the the package, as well as for the core to use from that
package. The most commonly used core reference is `arduino:arduino`
which works because both parts are identical, which is probably why this
bug has not shown up before.

This commit fixes the bug by simply using the right part to look up the
package.