Fix handling of referenced cores (#467)
* Add tests for ResolveFQBN

This adds tests for ResolveFQBN (which returns more details than the
FindBoardWithFQBN that was already tested), but also adds an extra
boards.txt with boards that are derived from another platform/core (i.e.
have `build.core=package:core`.

These new tests are currently failing, because of a bug and insufficient
error handling, which will be fixed in subsequent commits.

* Fix resolution of referenced cores

A core is referenced by specifying `build.core=package:core` in board
properties. However, the code used the second part (core) both for
looking up the the package, as well as for the core to use from that
package. The most commonly used core reference is `arduino:arduino`
which works because both parts are identical, which is probably why this
bug has not shown up before.

This commit fixes the bug by simply using the right part to look up the
package.

* Improve error message when referencing missing package

This used to say e.g.

	missing platform adafruit:referenced:avr required for build

where two packagenames (adafruit and referenced) were joined together
which makes no sense. Now, it just mentions the missing package, and the
fqbn that references it.

* Fix segfault when referenced platform is missing

Before, when a platform was referenced through a `build.core` and the
package was present by the platform/architecure was missing, a
nullpointer was passed to GetInstalledPlatformRelease, which would
segfault.

Now, a proper error message is returned.