[skip ci] Update docs re: platform.local.txt location
Previously, the location specified by the platform specification applied only to the folder structure of manually installed platforms. When a platform is installed via Boards Manager, there is an additional version folder level and it is under this folder the platform.local.txt must be placed.

The documentation now applies equally well to either platform installation type.