[skip changelog] Update outdated information in platform specification (#787)
* Update statement re: prevalence of usb to serial chip on Arduino boards

It's no longer true that "most" Arduino boards have a dedicated USB to serial chip (and becoming increasingly less so).

* Update documentation for sketch size regex properties

The platform specification was outdated in regard to the sketch size determination regular expression properties.

- Document recipe.size.regex.data
- Update Arduino AVR Boards' recipe.size.regex value

* Update docs re: platform.local.txt location

Previously, the location specified by the platform specification applied only to the folder structure of manually installed platforms. When a platform is installed via Boards Manager, there is an additional version folder level and it is under this folder the platform.local.txt must be placed.

The documentation now applies equally well to either platform installation type.

* Modernize statement re: upload tools

The statement was written at a time when there was only two Arduino platforms. Although still correct, it had a decidedly outdated feel that was fixed by a minor rewording.