[skip changelog] Fix issues with command reference section of the documentation's navigation pane (#801)
* Correct arduino-cli command name in command reference navigation

When no title is defined, the documentation system automatically generates a title from the filename. This resulted in the command being incorrectly called "Arduino cli" in the command reference section of the documentation's navigation pane.

* Add burn-bootloader to the command reference navigation

It's required to add an entry to mkdocs.yml before the command will be shown in the command reference section of the documentation's navigation pane.