Document how the `packager` value of a tool reference is defined
Referencing tools from another package is any important concept for platform authors to understand. Previously, you had to guess as to what the "packager" value of a tool reference is, which was not obvious due to it being defined outside the tool's definition, and thus not mentioned at all in the "Tools definitions" section of the specification, which is the obvious place to look for this information.

The situation is made more complicated by the fact that what is called "packager" in a tool reference is referred to as "vendor" everywhere else in the documentation.