Document that use of auto-set build.board property results in warning
The previous description made it sound like a convenience feature, with the build.board property only needed if you prefer a custom value for some reason. However, the resulting warning can cause confusion and act as a "red herring" for users of board definitions that don't define a build.board property.