[skip changelog] Use consistent indentation style in example snippets of package_index.json specification (#815)
* Use consistent index size for snippets in package_index.json specification

The predominant index size in the example snippets (as well as in Arduino's official package_index.json file) is two spaces. However, at some places in the example snippets indents of three or four spaces were used.

* Use true indentation level for package definition example snippet

All other example snippets in the documentation use the indentation level they will have in the full document, so for consistency the package definition snippet should do the same.