Use correct configuration for MkDocs pymdownx.magiclink extension (#823)
The pymdownx.magiclink Python-Markdown extension adds support for GitHub-style auto-linked issue/PR/commit/user references. In order for the issue/PR/commit references to work correctly in the case where no repository is specified, the extension must be configured with the owner and repository name of the repository that is using it. As a result of not having been updated after it was copied, the extension was previously configured for the squidfunk/mkdocs-material repository, which would result in any references that didn't specify a repository to be relative to that repository, rather than arduino/arduino-cli.