Format generated documentation
We now use Prettier to enforce a formatting standard on the Arduino CLI documentation. Previously, the generated documentation was not compliant with this standard, which caused some problems:

Incorrect list formatting of gRPC reference's table of contents. Prettier uses a two space indent. Because the stock Python-Markdown renderer requires a four space indent for nested lists, it was necessary to start using the mdx_truly_sane_lists extension to support the Prettier indent size. The use of the extension caused the four space indent used in the generated gRPC interface documentation's nested lists to no longer be rendered. It's possible that inconsistent formatting could cause similar issues. By making the formatting of all the documentation consistent, we can be sure that a specific renderer will handle all the docs correctly.

task docs:check fails when run on repository that contains generated docs. Although this could also have been fixed by excluding the generated documentation paths, the list formatting issue needed to be fixed anyway, and it happens to fix this issue as well.