Use v1 ref of arduino/setup-protoc action
The v1.1.0 version of the action neglected to define its repo-token input in the metadata, which results in a `Unexpected input(s) 'repo-token', valid inputs are ['version']` warning shown in the workflow summary on every run. The v1.1.1 release remedies this.

The use of the v1 ref will allow the workflows to benefit from any new releases of the action that don't cause breaking changes.