[skip changelog] Disable scheduled workflows that would always fail from running in forks (#888)
* Disable workflows that would always fail from running in forks

These workflows use resources that are only available when run in the arduino/arduino-cli repository, so would always fail when they run in a repository. They are of no value to fork repositories. They are triggered by the schedule event, so they cause regular annoying and confusing workflow failure notifications for every fork owner.

* Don't trigger link validation CI workflow on schedule event when in a fork