[skip changelog] Document the programmers system of Arduino platforms (#925)
* Document the programmers.txt configuration file of the Arduino platform system

This file is used to define external programmer configurations to be used for burning bootloaders and programming Arduino boards.

NOTE: the example programmer snippet is missing some of the properties that are defined in the real programmer definition in the Arduino AVR Boards platform. After a lot of investigation, experimentation, and even consulting with the original author of the code, I was still unable to find any use of these properties in the current or any past IDE version that supported the 1.5 platform format. If these properties have no purpose, they they only represent a source of confusion and maintenance burden for platform developers.

* Document the "Upload Using Programmer" configuration system of Arduino platforms

* Document the "Burn Bootloader" configuration system for Arduino platforms

* Document programmer inheritance from referenced core platforms

Previously, all programmers of all platforms of the same architecture could be used. The behavior of the Arduino IDE was changed in the 1.8.13 release in this matter, and Arduino CLI and Arduino Pro IDE followed the same behavior from the initial introduction of support for programmers. This means it's now important to document programmer inheritance.