[skip changelog]Fix contributing instructions re: skipping changelog (#916)
The arduino/arduino-cli repository's automatic changelog generation system adds a changelog item for every commit, except those that start with "[skip changelog]". This is useful for keeping the changelog to only mentioning the changes that are significant to the user.

All pull requests are merged using GitHub's "Squash & Merge" feature.

The behavior of "Squash & Merge":

- If the PR only contains a single commit, the pre-filled squash commit message is taken from the commit, not the PR.
- If the PR contains multiple commits, the prefilled squash commit title is the PR title.

This means that a PR will end up in the changelog against the wishes of the contributor when all of the following occur:

- The PR author followed the instructions and only added "[skip changelog]" to the PR title.
- The PR only contains a single commit.
- The maintainer who merges the PR uses the pre-filled commit message provided by GitHub.

The first two conditions are common. There is evidence that the third will also be common.

To make the instructions as easy as possible to follow, the contributor is now instructed to add "[skip changelog]" to both the commit message and PR title, even though it is really only necessary to add it to one or the other, depending on the conditions.