[skip changelog] Completely document resources not inherited from referenced platform (#940)
The inherited resources resulting from a core reference are clearly documented. However, the variant reference mentioned the non-inheritance of platform.txt properties alone. For me, this left it unclear whether the lack of mention of the other resources implied that variant and tool references resulted in their inheritance, and that it was only considered necessary to document it once in the platform reference documentation.

One option for resolving this would be to remove the statement that caused the confusion. However, since this is a complex subject, I thought it better to just explicitly document the non-inheritance of the resources in both the variant and tools reference sections of the documentation to make it completely clear.