Temporary: fix builds on dependent projects
Golang projects that uses the CLI as-a-library will fail to build
because the "replace" directive is valid only for the local project.
This should fix the build until https://github.com/GeertJohan/go.rice/pull/159
is merged.

Once the upstrea patch is merged this commit could be reverted and the
fork removed.