[skip changelog] Consolidate integration test instructions
Previously, instructions for running integration tests were present both in CONTRIBUTING.md as well as
in test/README.md.

Although most of the content was duplicated, there was
some important unique content on each page and the readers of one document would not necessarily know
of the existence of relevant information in the other.

Duplicate content is more work to maintain, and likely to end up becoming outdated.

So I have transfered the unique content from test/README.md to CONTRIBUTING.md and replaced it with a
link to that content.