[skip changelog] Update the tool definition example
The old tool definition example in the package_index.json specification was quite outdated. It did not demonstrate the
ARM hosts that have become more important since the specification was written. The macOS host value was antiquated.

This could lead to the reader being left with no clear idea of how to create the tool definitions for those hosts.