[skip changelog] Link to Sketch build process docs from Platform specification (#1025)
The "Build process" section of the Arduino platform specification contains a short summary of the Arduino sketch build
process. The "Sketch build process" page of the documentation provides much more detailed information, which wouldn't be
appropriate in the platform specification, but could well be of interest to a platform developer.

Adding a link to the "Sketch build process" documentation from that section of the Platform specification makes all
relevant information easily accessible to the reader.