Fix platform tool dependency determination (#1020)
* Fix platform tool dependency determination

The platform tool dependency check was giving false negatives due to comparing pointers instead of version values.

This caused tools to be removed during platform uninstallation even when another installed platform had a dependency on
that tool.

* Add integration test for tool removal during platform uninstall

Tool dependency by `arduino-cli core uninstall` is somewhat complex because it must only uninstall the tools that no
other platforms have a dependency on. If it doesn't, it breaks the other platform and the cause of this breakage would
likely not be obvious to the user.

So it's important to test to ensure this functionality continues to work correctly.