Fix issue with actions/checkout@v2
In certain workflows we use a custom Docker image for cross-compilation,
this image is old and unmaintained so it has an old version of git and
the latest version actions/checkout doesn't support it.
This causes the action to fallback to the Github API to download the
files but that doesn't create a git repository, thus the history is lost
but goreleaser needs it to generate the changelog so it fails.

So we use the older version of the action in certain cases to avoid
this failures.