[skip changelog] Make the command documentation of cache clean more clear
The terms "cache" and "clean" might bring compilation to mind, but the `arduino-cli cache clean` command has nothing to
do with compilation. Many users will likely not even be aware of the existence or purpose of the `directories.downloads`
folder. For this reason, it's important to make the purpose of this command clear.