[skip changelog] Use generally applicable name for package index specification
Previously the name of the specification was "package_index.json specification". Only the primary official package index
is allowed to be named package_index.json, but this specification applies to all package indexes. So "Arduino package
index specification" (shortened to "Package index specification" in the documentation navigaion pane) is less likely to
cause confusion and also easier to remember and talk about.