[skip changelog] Allow library names to start with number (#1136)
Previously, the Arduino library specification required the library.properties `name` field value to start with a letter.
Before Arduino IDE 1.8.4, library folder names starting with a number were not supported. Because the library folder for
Library Manager installations is named according to the library.properties name value, it was necessary to apply the
same restriction to the name value. It was determined that enough time has passed since the Arduino IDE 1.8.4 release to
reflect this change in the Arduino library specification.