Enhance lib install from git url and zip file
Previously installing from a zip file would extract the content directly
into the user library folder, this could have caused problem if the zip
file wasn't structured correctly.

To handle this we now extract to a temp directory to verify that the zip
is structured correctly and to infer the name of the library being
installed. If everything is fine we then copy it over to the user
library folder to finalize the installation.

Installing from a git repository has been enhanced too, after
cloning the repository to the user library folder we delete the .git
folder to make it a plain folder. Cloning should be faster too since we
now clone using depth of one commit to avoid downloading useless files.

Both when installing from a zip file or from a git repository we delete
the installed library folder if one with the same name is already
installed.