[skip changelog] Fix release process
The current release process uses git to get the tag at the current
commit. It might happen that are are multiple tags when calling that
command, in cases the rc and the final release are done on the same
commit, that would make the build process to fail.

This fixes that issue by taking only one tag.