Fix gRPC interface function to merge configs
This fix adds the possibility to set empty values with the Merge gRPC
function, previously they would have been ignored.

Because of this change I had also to modify the GetValue() function
since it would first check if the value was set using the
Viper.InConfig() function that wouldn't check for values set with
Viper.Set().