[skip changelog] Correct library+sketch specifications re: spaces in folder name
The previous mention of spaces as one of the characters allowed in library and sketch folder names was a copy/paste
error originating in the allowed characters in library.properties `name` properties. This specific part of the `name`
property specification doesn't apply to folder names. The reason it is allowed in the `name` property is because that
value is "sanitized" before being used by Library Manager for the library installation folder name by replacing all
spaces with underscores.