[skip changelog] Update Platform Specification link in warning (#1199)
When a boards platform uses an outdated `recipe.ar.pattern`, Arduino CLI displays a helpful message:

Unable to cache built core, please tell foo:bar@1.2.3 maintainers to follow http://goo.gl/QdCUjo

That short link points to the old Arduino Platform Specification page in the arduino/Arduino repository's wiki. The
platform specification has since been moved to the arduino/arduino-cli repository.

Even though the old page provides a
link to the new location, this message uses an anchor to direct the developer to the relevant section of the
specification. When they follow the link from the old page, they will simply end up at the top of the specification and
may have a hard time finding the information that tells them how to fix the platform. Updating the link restores the
intended behavior.