Exit with error when a sketch file is unreadable
When any sketch file that would normally be compiled is unreadable,
SketchLoad would previously just silently skip it, leading to
potentially unexpected results. It is probably better to just error out
and let the user sort out the problem instead.

Additionally, this prepares for improving the handling of broken
symlinks in the next commit.