[skip changelog] Remove outdated information re: debug config from platform spec
Since the time the documentation for the initial experimental debugging configuration system was added to the Arduino
Platform Specification, the system has changed significantly, resulting in some of the documentation no longer being
applicable to the current version of Arduino CLI. This only removes the outdated information, leaving the parts that are
still correct. The new debug system will be fully documented at a later time.