Redirect pre-versioning docs website URLs to latest (#1202)
The documentation website was originally unversioned. The outdated content from that time is still present in the root
of the website. The only redirect that was set up is the home page, but there are links around the Internet to the other
pages. People visiting the website from those links are getting content that is outdated by nearly a year, during which
time some significant changes have been made.

So it's necessary to set up redirects from each of the non-versioned pages to its counterpart in the "latest" version.