Improved lib detection: check for matching name in library.properties
The library may be stored in a directory that doesn't match the library
name, for example we had a case in the wild where the directories:

   libraries/onewire_2_3_4/...
   libraries/onewireng_1_2_3/...

were used instead of:

   libraries/OneWire/...
   libraries/OneWireNg/...

this lead to incorrect selection of onewireng_1_2_3 when using OneWire.h
(because the OneWireNg had an architecture=avr that had priority over
the architecture=* of onewire_2_3_4).

This commit will restore priority straight.